+++ RELEASE-NOTES.md
@@ -17,6 +17,7 @@
 * #623 ECharts: VisualMap seriesIndex mapping
 * #624 QL: "sqrt(..)" expression
 * #625 Upgrade jackson to 2.19.4
+* #627 NumExp.median() and quantile() default implementations are not using filter argument 
 
 ## Release 2.0.0-M6